requirements.txt
bcrypt==4.1.3
blinker==1.8.2
click==8.1.7
flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
flask-sqlalchemy==3.1.1
flask-wtf==1.2.1
greenlet==3.0.3
importlib-metadata==7.1.0
itsdangerous==2.2.0
jinja2==3.1.4
MarkupSafe==2.1.5
mysqlclient==2.2.4
PyMySQL==1.1.1
python-dotenv==1.0.1
SQLAlchemy==2.0.30
typing-extensions==4.11.0
werkzeug==3.0.3
wtforms==3.1.2
zipp==3.18.2
